Move WebhookThread to its own file
