+++ full-service/src/service/sync.rs
@@ -30,12 +30,6 @@ use mc_transaction_core::{
 };
 use rayon::prelude::*;
 
-use reqwest::Url;
-use reqwest::{
-    blocking::Client, // FIXME: if we don't want this to be blocking, we'll be awaiting futures
-    header::{HeaderMap, HeaderValue, CONTENT_TYPE},
-};
-use serde_json::json;
 use std::{
     collections::HashMap,
     convert::TryFrom,
@@ -56,129 +50,6 @@ pub struct SyncThread {
 
     /// Stop trigger, used to signal the thread to terminate.
     stop_requested: Arc<AtomicBool>,
-}
-
-// FIXME: refactor this to its own file - it's gotten big enough
-pub struct WebhookThread {
-    /// The main sync thread handle.
-    join_handle: Option<thread::JoinHandle<()>>,
-
-    /// Stop trigger, used to signal the thread to terminate.
-    stop_requested: Arc<AtomicBool>,
-}
-
-impl WebhookThread {
-    pub fn start(
-        server_url: Url,
-        accounts_with_deposits: Arc<Mutex<HashMap<AccountID, bool>>>,
-        restart: Arc<Mutex<AtomicBool>>,
-        logger: Logger,
-    ) -> Self {
-        // Start the webhook thread.
-
-        let stop_requested = Arc::new(AtomicBool::new(false));
-        let thread_stop_requested = stop_requested.clone();
-
-        // Question: Should we consider only spawning a thread when there
-        // have been received txos, and therefore something to send?
-        // Answer: For now, we are ok with having this thread running all the time,
-        // because it is lightweight enough. If we find that it is causing issues,
-        // we will revisit. The solution would be to make this async, or to only
-        // spawn the thread when there are txos to send. There may be an advantage to
-        // leaving the connection open.
-
-        let join_handle = Some(
-            thread::Builder::new()
-                .name("webhook".to_string())
-                .spawn(move || {
-                    log::debug!(logger, "Webhook thread started.");
-
-                    let client = Client::builder()
-                        .build()
-                        .expect("Failed creating reqwest client");
-                    let mut json_headers = HeaderMap::new();
-                    json_headers.insert(CONTENT_TYPE, HeaderValue::from_static("application/json"));
-
-                    loop {
-                        if thread_stop_requested.load(Ordering::SeqCst) {
-                            log::debug!(logger, "WebhookThread stop requested.");
-                            break;
-                        }
-                        // Gather the current accounts_to_send, then wipe the contents
-                        let accounts_to_send: Vec<_> = accounts_with_deposits
-                            .lock()
-                            .unwrap()
-                            .clone()
-                            .iter()
-                            .filter(|&(_k, &v)| v == true)
-                            .map(|(k, _v)| k.clone())
-                            .collect();
-
-                        // Delete the keys that we're alerting on
-                        for key in accounts_to_send.iter() {
-                            log::debug!(logger, "Account to send: {:?}", key);
-                            accounts_with_deposits.lock().unwrap().remove(&key);
-                        }
-
-                        // Only read restart, do not set
-                        let restart_to_send = restart.lock().unwrap().load(Ordering::Relaxed);
-
-                        if accounts_to_send.len() > 0 {
-                            log::debug!(logger, "@@@@ ABOUT TO SEND THE WEBHOOK");
-                            // Question: will this keep the connection open? Or will it
-                            // close the connection after this request?
-                            match client
-                                .post(server_url.clone())
-                                .body(
-                                    json!(
-                                        {
-                                            "accounts": accounts_to_send,
-                                            "restart": restart_to_send,
-                                        }
-                                    )
-                                    .to_string(),
-                                )
-                                .send()
-                            {
-                                Ok(response) => match response.error_for_status() {
-                                    Ok(_) => (),
-                                    Err(e) => {
-                                        log::error!(
-                                            logger,
-                                            "Failed getting webhook response: {:?}",
-                                            e
-                                        );
-                                    }
-                                },
-                                Err(e) => {
-                                    log::error!(logger, "Failed sending webhook request: {:?}", e);
-                                }
-                            }
-                        }
-                        // FIXME: this duration should be as often as we are polling
-                        // for new blocks from consensus
-                        thread::sleep(Duration::from_millis(10));
-                    }
-                })
-                .expect("failed starting webhook thread"),
-        );
-        Self {
-            join_handle,
-            stop_requested,
-        }
-    }
-    pub fn stop(&mut self) {
-        self.stop_requested.store(true, Ordering::SeqCst);
-        if let Some(join_handle) = self.join_handle.take() {
-            join_handle.join().expect("WebhookThread join failed");
-        }
-    }
-}
-
-impl Drop for WebhookThread {
-    fn drop(&mut self) {
-        self.stop();
-    }
 }
 
 impl SyncThread {
